fix loading

+++ pcv/lib/views/home/about/profile.dart
@@ -12,122 +12,75 @@ class ProfileScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: appBarMethod(title: "Profile"),
-      body: SafeArea(
-        child: SingleChildScrollView(
-          child: Padding(
-            padding: const EdgeInsets.symmetric(horizontal: 15),
-            child: Column(
-              children: [
-                BlocConsumer<AboutBloc, AboutState>(
-                  listener: (context, state) {
-                    if (state is ErrorAboutState) {
-                      ScaffoldMessenger.of(context)
-                          .showSnackBar(SnackBar(content: Text(state.msg)));
-                    }
-                    if (state is SuccessAboutState) {
-                      showDialog(
-                          context: context,
-                          builder: (context) => const Center(
-                                child: Text("Complate"),
-                              ));
-                      Future.delayed(const Duration(seconds: 3), () {
-                        Navigator.of(context).pop();
-                      });
-                    }
-                  },
-                  builder: (context, state) {
-                    return BlocBuilder<AboutBloc, AboutState>(
-                      builder: (context, state) {
-                        if (state is GetAboutState) {
+    return GestureDetector(
+      onTap: () => FocusScope.of(context).unfocus(),
+      child: Scaffold(
+        appBar: appBarMethod(title: "Profile"),
+        body: SafeArea(
+          child: SingleChildScrollView(
+            child: Padding(
+              padding: const EdgeInsets.symmetric(horizontal: 15),
+              child: Column(
+                children: [
+                  BlocConsumer<AboutBloc, AboutState>(
+                    listener: (context, state) {
+                      if (state is ErrorAboutState) {
+                        ScaffoldMessenger.of(context)
+                            .showSnackBar(SnackBar(content: Text(state.msg)));
+                      }
+                      if (state is SuccessAboutState) {
+                        showDialog(
+                            context: context,
+                            builder: (context) => const Center(
+                                  child: Text("Complate"),
+                                ));
+                        Future.delayed(const Duration(seconds: 3), () {
+                          Navigator.of(context).pop();
+                        });
+                      }
+                    },
+                    builder: (context, state) {
+                      return BlocBuilder<AboutBloc, AboutState>(
+                        builder: (context, state) {
+                          if (state is GetAboutState) {
+                            return ProfileWidget(
+                              birthdayController: TextEditingController(
+                                  text: state.about.data?.birthday),
+                              usernameController: TextEditingController(
+                                  text: state.about.data?.name),
+                              titPoController: TextEditingController(
+                                  text: state.about.data?.titlePosition),
+                              locationController: TextEditingController(
+                                  text: state.about.data?.location),
+                              phoneController: TextEditingController(
+                                  text: state.about.data?.phone),
+                              aboutController: TextEditingController(
+                                  text: state.about.data?.about),
+                            );
+                          }
                           return ProfileWidget(
-                            birthdayController: TextEditingController(
-                                text: state.about.data?.birthday),
-                            usernameController: TextEditingController(
-                                text: state.about.data?.name),
-                            titPoController: TextEditingController(
-                                text: state.about.data?.titlePosition),
-                            locationController: TextEditingController(
-                                text: state.about.data?.location),
-                            phoneController: TextEditingController(
-                                text: state.about.data?.phone),
-                            aboutController: TextEditingController(
-                                text: state.about.data?.about),
+                            birthdayController: TextEditingController(),
+                            usernameController: TextEditingController(),
+                            titPoController: TextEditingController(),
+                            locationController: TextEditingController(),
+                            phoneController: TextEditingController(),
+                            aboutController: TextEditingController(),
                           );
-                        }
-                        return ProfileWidget(
-                          birthdayController: TextEditingController(),
-                          usernameController: TextEditingController(),
-                          titPoController: TextEditingController(),
-                          locationController: TextEditingController(),
-                          phoneController: TextEditingController(),
-                          aboutController: TextEditingController(),
-                        );
-                      },
-                    );
-                  },
-                ),
-                ButtonWidget(
-                  onPressed: () {
-                    showDialog(
-                        context: context,
-                        builder: (context) => AlertDialog(
-                              title: const Text(
-                                "LogOut",
-                                style: TextStyle(color: Colors.red),
-                              ),
-                              content: const Text("Are You Sure want Logout?"),
-                              actions: [
-                                TextButton(
-                                    onPressed: () {
-                                      Navigator.of(context).pop();
-                                    },
-                                    child: const Text(
-                                      "CANCEL",
-                                    )),
-                                TextButton(
-                                    onPressed: () {
-                                      prefs?.clear();
-                                      Navigator.pushAndRemoveUntil(
-                                          context,
-                                          MaterialPageRoute(
-                                              builder: (context) =>
-                                                  const SignInScreen()),
-                                          (route) => false);
-                                    },
-                                    child: const Text("Logout",
-                                        style: TextStyle(color: Colors.red)))
-                              ],
-                            ));
-                  },
-                  text: "Logout",
-                  color: Colors.red,
-                ),
-                BlocListener<AboutBloc, AboutState>(
-                  listener: (context, state) {
-                    if (state is SuccessDeleteAccount) {
-                      Navigator.pushAndRemoveUntil(
-                          context,
-                          MaterialPageRoute(
-                              builder: (context) => const SignInScreen()),
-                          (route) => false);
-                    } else if (state is ErrorAboutState) {
-                      ScaffoldMessenger.of(context)
-                          .showSnackBar(SnackBar(content: Text(state.msg)));
-                    }
-                  },
-                  child: ButtonWidget(
+                        },
+                      );
+                    },
+                  ),
+                  ButtonWidget(
                     onPressed: () {
                       showDialog(
                           context: context,
                           builder: (context) => AlertDialog(
                                 title: const Text(
-                                  "DELETE ACCOUNT",
+                                  "LogOut",
                                   style: TextStyle(color: Colors.red),
                                 ),
                                 content:
-                                    const Text("Are You Sure Delete Account"),
+                                    const Text("Are You Sure want Logout?"),
                                 actions: [
                                   TextButton(
                                       onPressed: () {
@@ -138,20 +91,73 @@ class ProfileScreen extends StatelessWidget {
                                       )),
                                   TextButton(
                                       onPressed: () {
-                                        context
-                                            .read<AboutBloc>()
-                                            .add(DeleteAccount());
+                                        prefs?.clear();
+                                        Navigator.pushAndRemoveUntil(
+                                            context,
+                                            MaterialPageRoute(
+                                                builder: (context) =>
+                                                    const SignInScreen()),
+                                            (route) => false);
                                       },
-                                      child: const Text("DELETE",
+                                      child: const Text("Logout",
                                           style: TextStyle(color: Colors.red)))
                                 ],
                               ));
                     },
-                    text: "DELETE ACCOUNT",
+                    text: "Logout",
                     color: Colors.red,
                   ),
-                ),
-              ],
+                  BlocListener<AboutBloc, AboutState>(
+                    listener: (context, state) {
+                      if (state is SuccessDeleteAccount) {
+                        prefs?.clear();
+                        Navigator.pushAndRemoveUntil(
+                            context,
+                            MaterialPageRoute(
+                                builder: (context) => const SignInScreen()),
+                            (route) => false);
+                      } else if (state is ErrorAboutState) {
+                        ScaffoldMessenger.of(context)
+                            .showSnackBar(SnackBar(content: Text(state.msg)));
+                      }
+                    },
+                    child: ButtonWidget(
+                      onPressed: () {
+                        showDialog(
+                            context: context,
+                            builder: (context) => AlertDialog(
+                                  title: const Text(
+                                    "DELETE ACCOUNT",
+                                    style: TextStyle(color: Colors.red),
+                                  ),
+                                  content:
+                                      const Text("Are You Sure Delete Account"),
+                                  actions: [
+                                    TextButton(
+                                        onPressed: () {
+                                          Navigator.of(context).pop();
+                                        },
+                                        child: const Text(
+                                          "CANCEL",
+                                        )),
+                                    TextButton(
+                                        onPressed: () {
+                                          context
+                                              .read<AboutBloc>()
+                                              .add(DeleteAccount());
+                                        },
+                                        child: const Text("DELETE",
+                                            style:
+                                                TextStyle(color: Colors.red)))
+                                  ],
+                                ));
+                      },
+                      text: "DELETE ACCOUNT",
+                      color: Colors.red,
+                    ),
+                  ),
+                ],
+              ),
             ),
           ),
         ),
